[FEAT] constants 공통 폴더로 분리
