--- src/app/constants/chart.ts
@@ -1,7 +1,0 @@
-export const CHART_COLORS = {
-  chart1: '#80FFA5',
-  chart2: '#00DDFF',
-  chart3: '#37A2FF',
-  chart4: '#FF0087',
-  chart5: '#FFBF00'
-};
