+++ src/constants/index.ts
@@ -1,0 +1,4 @@
+import * as ChartOptions from './options/chart';
+import * as SelectOptions from './options/select';
+
+export { ChartOptions, SelectOptions };
