+++ src/constants/options/select.ts
@@ -1,4 +1,5 @@
-export const SELECT_FILTER = [
+// select
+export const MOCK_SELECT_OPTIONS = [
   { value: '1', label: '1년' },
   { value: '3', label: '3년' },
   { value: '5', label: '5년' },
